+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Improvement:
 Bug fix:
 - [#1075](https://github.com/elysiajs/elysia/issues/1075) nested async plugins mismatch routes to handlers
 - [#1070](https://github.com/elysiajs/elysia/issues/1070) .mount is mutating the incoming request method
+- [#1073](https://github.com/elysiajs/elysia/issues/1073) file type validation not working
 
 # 1.2.15 - 19 Feb 2025
 Bug fix: